+++ elucidate-server/src/main/java/com/digirati/elucidate/repository/AnnotationSearchRepository.java
@@ -17,7 +17,7 @@ public interface AnnotationSearchRepository {
 
     List<W3CAnnotation> getAnnotationsByTemporal(boolean searchAnnotations, boolean searchBodies, boolean searchTargets, String[] types, Date since);
 
-    List<W3CAnnotation> getAnnotationsByRange(String targetId, int rangeStart, int rangeEnd);
+    List<W3CAnnotation> getAnnotationsByRange(String targetId, float rangeStart, float rangeEnd);
 
-    List<W3CAnnotation> getAnnotationsByOverlap(String targetId, int rangeStart, int rangeEnd);
+    List<W3CAnnotation> getAnnotationsByOverlap(String targetId, float rangeStart, float rangeEnd);
 }
